+++ .env
@@ -1,2 +1,3 @@
 VITE_SUPABASE_URL=https://zakuvvzlgkswrefebmsg.supabase.co/
 VITE_SUPABASE_ANON_KEY=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6Inpha3V2dnpsZ2tzd3JlZmVibXNnIiwicm9sZSI6ImFub24iLCJpYXQiOjE3MjQxMzgyNDYsImV4cCI6MjAzOTcxNDI0Nn0.9fjxmFpfMBhtKEiEFOnl6kKhyMcchG4WfiAP98867OU
+VITE_MAPBOX_KEY=pk.eyJ1IjoiZmFubnlrYXJsc3NvbiIsImEiOiJja2d3Y2Z1YjgwOHVmMnNwOGp0MDJhdnZqIn0.UAHwkHlBKkQqKK3hJw0GXA
