+++ src/Components/ViewNotes/ViewNotes.module.css
@@ -1,0 +1,35 @@
+.notePopup {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0, 0, 0, 0.8);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    z-index: 1000;
+}
+  
+.notePopupContent {
+background-color: white;
+padding: 20px;
+border-radius: 8px;
+max-width: 80%;
+max-height: 80%;
+overflow-y: auto;
+box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
+}
+
+.notePopupContentButton {
+position: absolute;
+top: 10px;
+right: 10px;
+background-color: red;
+color: white;
+border: none;
+padding: 10px;
+cursor: pointer;
+border-radius: 5px;
+}
+
